Support installation folders other than in Program Files
At the moment, VS installations must recide in the Program Files (x86) folder. If it's installed in another location, then the path isn't resolved anymore. This uses vswhere, which is always installed on the same location), and uses the given paths. This now correctly resolves a path on another drive.