Dot-source Invoke-Build in ib.cmd to make Invoke-Build available in closures
In the end, looks like dot-sourcing scripts and then calling functions
from them is prefereded approach in PowerShell. As an example this is
how PowerShell profile scripts work (https://stackoverflow.com/a/36633913/1334364),
and how the `-File` argument to powershell.exe behaves.

Without this change build script like below will crash with "ERROR: The term 'exec' is not recognized as the name of a cmdlet, function, script file, or operable program.":

```powershell
@("Alice", "Bob", "Charlie") | ForEach-Object {
  $taskName = $PSItem
  task "Hello$taskName" {
    exec { cmd /c "echo Hello $taskName" }
    # GetNewClosure() is required to mimic real closures in PowerShell
  }.GetNewClosure()
}
```

And run like: ib HelloBob

Also using `-File` instead of `Command` would fix this issue, as it also
dot-source the script before executing it. But with `-File` we loos
ability to pass complex scripts as parameters (thing hashtables) when
calling ib.cmd from Command Prompt.